+++ packages/ricos-editor/web/src/index.ts
@@ -1,8 +1,2 @@
 export * from 'ricos-common';
-
-import RicosEditorSwitcher from './RicosEditorSwitcher';
-
-export default RicosEditorSwitcher;
-export { RicosEditorSwitcher as RicosEditor };
-
-export { RicosEditor as RicosEditorType } from './RicosEditor';
+export { default as RicosEditor, RicosEditor as RicosEditorType } from './RicosEditor';